+++ rules.js
@@ -1,0 +1,5 @@
+const rules = require('./lib/configs/rules.json')
+
+module.exports = {
+  rules: { ...rules },
+}