+++ test/index.js
@@ -1,0 +1,8 @@
+const foo = {
+  bar: 1,
+  baz: 2,
+};
+
+console.log('foo');
+
+module.exports = foo;